Enable hosted control plane by default UI work
