+++ frontend/public/locales/en/translation.json
@@ -1442,7 +1442,6 @@
   "Hosted cluster requires a manual import.": "Hosted cluster requires a manual import.",
   "Hosted control plane": "Hosted control plane",
   "Hosted control plane is already upgrade to the latest version available. Cluster node pools can be upgraded to match the control plane.": "Hosted control plane is already upgrade to the latest version available. Cluster node pools can be upgraded to match the control plane.",
-  "Hosted control plane operator and hypershift add-on must be enabled in order to continue": "Hosted control plane operator and hypershift add-on must be enabled in order to continue",
   "Hosted control plane operator must be enabled in order to continue": "Hosted control plane operator must be enabled in order to continue",
   "Hosted Control Planes for OpenShift Virtualization are created using the KubeVirt platform type. To create the Hosted Control Plane, copy and paste the following command: ": "Hosted Control Planes for OpenShift Virtualization are created using the KubeVirt platform type. To create the Hosted Control Plane, copy and paste the following command: ",
   "hostInfraEnvCounts": "$t(hostCount, {\"count\": {{hostCount}} }) in $t(infraEnvCount, {\"count\": {{infraEnvCount}} })",
@@ -2558,7 +2557,6 @@
   "table.userManagement": "User management",
   "table.zone": "Zone",
   "Technical community": "Technical community",
-  "Technology preview": "Technology preview",
   "Template": "Template",
   "Template details": "Template details",
   "Template Type": "Template type",
